+++ CHANGELOG.md
@@ -96,6 +96,7 @@ and the versioning aim to respect [Semantic Versioning](http://semver.org/spec/v
   default_scalars.csv instead of default_costs_efficiencies.csv
 - Default variable_costs are passed with input_parameters for storages
 - Pass time series instead of scalar with efficiency for central heat pump
+- Fix wind+pv_ground default values in panel settings
 
 ### Removed